aes.py
AES python implementation
The pad method recieves plaintext and adds bytes to make it of block size 128 bits while the unpad method removes all the extra added characters.
In the encrypt method, we receive the plaintext and pad it in order to encrypt it. An iv is generated with block size of 128 bits. The cipher is then generated using iv and plaintext using the CBC mode.