Create dsa.py
Python code for digital signature 
The function keys generate the private key of length 1024 and the public key. The function for encryption takes the public key and plain text as arguments and returns an encoded string of ciphertext. The decryption function takes ciphertext and private key as arguments and returns the plaintext. 
The digital signature provides user authentication.